requirements.txt
pyqt6
python-can
cantools
pyinstaller